modified:   C_test_project/5_loop_statement/practice2/a.exe 	modified:   C_test_project/5_loop_statement/practice2/practice.c

+++ C_test_project/5_loop_statement/practice2/practice.c
@@ -8,12 +8,12 @@
 #include <math.h>
 int Sum_fraction()
 {
-    float sum=0;
+    double sum=0;
     for(int i=1;i<=100;i++)
     {
         sum=sum+(1/(pow(-1,i-1)*i));
     }
-    printf("%f\n",sum);
+    printf("%lf\n",sum);
     return 0;
 }
 int main()
